flask/segmentation/requirements.txt
addict==2.4.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.11.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyter-packaging==0.12.0
jupyter-server==1.16.0
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
markdown==3.3.6
markupsafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.2.0
open3d==0.15.2
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.1.0
pip==22.0.4
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
protobuf==3.20.*
pure-eval==0.2.2
pycparser==2.21
pygments==2.11.2
pyparsing==3.0.7
pyquaternion==0.9.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyyaml==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
send2trash==1.8.0
setuptools==61.3.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.8.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tomlkit==0.10.1
https://open3d-downloads.b-cdn.net/torch-1.8.2-cp38-cp38-linux_x86_64.whl ; python_version == '3.8'
https://open3d-downloads.b-cdn.net/torch-1.8.2-cp39-cp39-linux_x86_64.whl ; python_version == '3.9'
-f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torchvision==0.9.2+cu111
tensorboard
tensorflow~=2.5.2
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.0
zipp==3.7.0